+++ R/Rraven-package.R
@@ -4,7 +4,7 @@
 #' between R and 'Raven' sound analysis software (Cornell Lab of Ornithology)
 #' @details License: GPL (>= 2)  
 #' 
-#'  @section Functions:
+#' @section Functions:
 #'   
 #'   \code{\link{exp_raven}}: Export R selection tables into 'Raven' selection file format
 #'  
@@ -12,7 +12,7 @@
 #'
 #'   \code{\link{exp_est}}: Export wave objects of extended selection tables as sound files
 #'   
-#'   \code{\link{extract_ts}}: Extract time series parameters from data imported from Raven
+#'   \code{\link{extract_ts}}: Extract time series parameters from data imported froNm Raven
 #'   
 #'   \code{\link{fix_path}}: Modify sound file path in Raven's selection tables 
 #'   
@@ -33,8 +33,9 @@
 #'   \code{\link{to_sound_selection}}: Convert Raven's selection files into sound selection files
 #'      
 #' @import warbleR 
-#' @import NatureSounds 
 #' @import pbapply
+#' @import methods
+#' @importFrom methods is
 #' @importFrom seewave duration pastew
 #' @importFrom tuneR normalize writeWave
 #' @importFrom stats approx as.ts