+++ R/exp_raven.R
@@ -3,12 +3,12 @@
 #' \code{exp_raven} exports selection tables as 'Raven' selection data in .txt format.
 #' @usage exp_raven(X, path = NULL, file.name = NULL, khz.to.hz = TRUE, 
 #' sound.file.path = NULL, single.file = TRUE, parallel = 1, pb = TRUE)
-#' @param X Object of class data frame or \code{\link[warbleR]{selection_table}} containing columns for sound file (sound.files), selection (selec), start and end time of signals ('start' and 'end') and low and high frequency ('bottom.freq' and 'top.freq', optional). See example data 'selec_table' in the \code{\link{warbleR}}) package.
+#' @param X Object of class data frame or \code{\link[warbleR]{selection_table}} containing columns for sound file (sound.files), selection (selec), start and end time of signals ('start' and 'end') and low and high frequency ('bottom.freq' and 'top.freq', optional). See example data 'selec_table' in the \href{https://cran.r-project.org/package=warbleR}{warbleR}) package.
 #' @param path A character string indicating the path of the directory in which to save the selection files.
 #' If not provided (default) the function saves the file into the current working directory.
 #' @param file.name Name of the output .txt file. If \code{NULL} then the sound file names are used instead. If multiple
 #' selection files are generated (see 'single.file') then the sound files names are added to the provided 'file.name'. Ignored if \code{single.file = FALSE}.
-#' @param khz.to.hz Logical. Controls if frequency variables should be converted from kHz (the unit used by other bioacoustic analysis R packages like \code{\link{warbleR}}) to Hz (the unit used by Raven). Default is \code{TRUE}.
+#' @param khz.to.hz Logical. Controls if frequency variables should be converted from kHz (the unit used by other bioacoustic analysis R packages like \href{https://cran.r-project.org/package=warbleR}{warbleR}) to Hz (the unit used by Raven). Default is \code{TRUE}.
 #' @param sound.file.path A character string indicating the path of the 
 #' directory containing the sound file(s). Providing this information allows
 #'  to open both sound file and selection table simultaneously. This can be
@@ -28,7 +28,7 @@
 #' export them as a single selection file or as multiple selection files (one for each sound file). 
 #' No objects are returned in the R environment.
 #' @details The function exports selection tables (as the ones used in the R 
-#' package \code{\link{warbleR}}) into the 'Raven' selection file format ('.txt').
+#' package \href{https://cran.r-project.org/package=warbleR}{warbleR}) into the 'Raven' selection file format ('.txt').
 #'  This can be useful to obtain additional Raven
 #' measurements on existing selections by adding new measurements to the 
 #' selection table once in Raven. Note that selection labels must be numeric and non-duplicated 
@@ -69,21 +69,21 @@ exp_raven <- function(X, path = NULL, file.name = NULL, khz.to.hz = TRUE, sound.
   
   #check path to working directory
   if (is.null(path)) path <- getwd() else 
-    if (!dir.exists(path)) stop("'path' provided does not exist") else
+    if (!dir.exists(path)) stop2("'path' provided does not exist") else
       path <- normalizePath(path)
   
   #if X is not a data frame
-  if (!any(is.data.frame(X), warbleR::is_selection_table(X))) stop("X is not of a class 'data.frame', 'selection_table'")
+  if (!any(is.data.frame(X), warbleR::is_selection_table(X))) stop2("X is not of a class 'data.frame', 'selection_table'")
   
   if (warbleR::is_selection_table(X)) X <- as.data.frame(X)
   
   if (!all(c("sound.files", "selec", 
              "start", "end") %in% colnames(X))) 
-    stop(paste(paste(c("sound.files", "selec", "start", "end")[!(c("sound.files", "selec", 
+    stop2(paste(paste(c("sound.files", "selec", "start", "end")[!(c("sound.files", "selec", 
                                                                    "start", "end") %in% colnames(X))], collapse=", "), "column(s) not found in data frame"))
   
   #stop if more than 1 sound file is found in X
-  if (length(unique(X$sound.files)) > 1 & is.null(sound.file.path)) stop("'sound.file.path' must be provided when including selections from multiple sound files")
+  if (length(unique(X$sound.files)) > 1 & is.null(sound.file.path)) stop2("'sound.file.path' must be provided when including selections from multiple sound files")
   
   if (length(unique(X$sound.files)) == 1) single.file <- TRUE else
     if(!single.file) file.name <- NULL
@@ -95,12 +95,12 @@ exp_raven <- function(X, path = NULL, file.name = NULL, khz.to.hz = TRUE, sound.
     #count number of sound files in working directory and if 0 stop
     recs.wd <- list.files(path = sound.file.path, pattern = "\\.wav$|\\.aif$|\\.flac$|\\.mp3$", ignore.case = TRUE)
     if (!all(unique(X$sound.files) %in% recs.wd)) 
-      stop("Some (or all) sound files are not in the working directory")
+      stop2("Some (or all) sound files are not in the working directory")
   }
   
-  if (any(is.na(X$start), is.na(X$end))) stop("'NAs' found in start and/or end column")
+  if (any(is.na(X$start), is.na(X$end))) stop2("'NAs' found in start and/or end column")
   
-  if (any(!is.numeric(X$start), !is.numeric(X$end))) stop("start and/or end column(s) are not numeric")
+  if (any(!is.numeric(X$start), !is.numeric(X$end))) stop2("start and/or end column(s) are not numeric")
   
   # convert to Hz
   if ("bottom.freq" %in% names(X) & khz.to.hz)
@@ -133,7 +133,7 @@ exp_raven <- function(X, path = NULL, file.name = NULL, khz.to.hz = TRUE, sound.
     
     if (length(unique(X$'Begin File')) > 1 & single.file)
     {
-      durs <- warbleR::wavdur(path = sound.file.path, files = as.character(X$'Begin File'[!duplicated(X$'Begin File')]))
+      durs <- warbleR::duration_sound_files(path = sound.file.path, files = as.character(X$'Begin File'[!duplicated(X$'Begin File')]))
       durs$cumdur <- cumsum(durs$duration)
      durs <- durs[durs$sound.files %in% X$'Begin File', ]
     
@@ -173,14 +173,11 @@ if (single.file | nrow(X) == 1)
   rvn_msr_nms <-  c("Begin Time (s)", "End Time (s)", "Low Frequency (Hz)", "High Frequency (Hz)", "File Offset (s)", "Begin File", "Begin Path", "1st Quartile Frequency (Hz)", "1st Quartile Frequency Relative", "1st Quartile Time (s)", "1st Quartile Time Relative", "3rd Quartile Frequency (Hz)", "3rd Quartile Frequency Relative", "3rd Quartile Time (s)", "3rd Quartile Time Relative", "Aggregate Entropy (bits)", "Average Amplitude (u)", "Average Entropy (bits)", "Average Power (dB)", "Bandwidth 90% (Hz)", "Begin Clock Time", "Begin Date", "Begin Date Time", "Begin File Sample (samples)", "Begin Hour", "Begin Sample (samples)", "Center Frequency (Hz)", "Center Time (s)", "Center Time Relative", " Delta Frequency (Hz)", "Delta Power (dB)", "Delta Time (s)", "Duration 90% (s)", "End Clock Time", "End Date", "End File", "End File Sample (samples)",
                     "End Path", "End Sample (samples)", "Energy (dB)", "Filtered RMS Amplitude (U)", "Frequency 5% (Hz)", "Frequency 5% Relative", "Frequency 95% (Hz)", "Frequency 95% Relative", "Frequency Contour Percentile  5% (Hz)", "Frequency Contour Percentile 25% (Hz)", "Frequency Contour Percentile 50% (Hz)", "Frequency Contour Percentile 75% (Hz)", "Frequency Contour Percentile 95% (Hz)", "IQR Bandwidth (Hz)", "IQR Duration (s)", "Inband Power (dB)", "Length (frames)", "Leq (dB)", "Max Amplitude (U)", "Max Bearing (deg)", "Max Entropy (bits)", "Max Frequency (Hz)", "Max Power (dB)", "Max Time (s)", "Min Amplitude (U)", "Min Entropy (bits)", "Min Time (s)", "Peak Amplitude (U)", "Peak Correlation (U)", "Peak Frequency (Hz)", "Peak Frequency Contour (Hz)", "Peak Frequency Contour Average Slope (Hz/ms)", "Peak Frequency Contour Max Frequency (Hz)", "Peak Frequency Contour Max Slope (Hz/ms)", "Peak Frequency Contour Min Frequency (Hz)", "Peak Frequency Contour Min Slope (Hz/ms)", "Peak Frequency Contour Number of Inflection Points", "Peak Frequency Contour Slope (Hz/ms)", "Peak Lag (s)", "Peak Power (dB)", "Peak Time (s)", "Peak Time Relative", "RMS Amplitude (U)", "SEL (dB)", "Sample Length (samples)", "Time 5% (s)", "Time 5% Relative", "Time 95% (s)", "Time 95% Relative")
   
-  # set pb options 
-  pbapply::pboptions(type = ifelse(pb, "timer", "none"))
-  
   # set clusters for windows OS
   if (Sys.info()[1] == "Windows" & parallel > 1)
     cl <- parallel::makePSOCKcluster(getOption("cl.cores", parallel)) else cl <- parallel  
   
-  out <- pbapply::pblapply(seq_len(nrow(row.list)), cl = cl, function(x){
+  out <- warbleR:::.pblapply(pbar = pb, X = seq_len(nrow(row.list)), cl = cl,  message = "exporting annotations", total = 1, function(x){
   
   if (is.null(file.name)) file.name2 <- "" else file.name2 <- file.name