+++ R/match_wav_case.R
@@ -3,7 +3,7 @@
 #' \code{match_wav_case} fixes the extension case of sound files in a selection table.
 #' @usage match_wav_case(X, path = NULL, output = "data.frame", verbose = TRUE)
 #' @param X Data frame containing columns for sound file (sound.files) and selection
-#'  (selec). See example data 'lbh_selec_table' in the \code{\link{warbleR}}) package.
+#'  (selec). See example data 'lbh_selec_table' in the \href{https://cran.r-project.org/package=warbleR}{warbleR}) package.
 #' @param path A character string indicating the path of the directory in which to look for sound files. 
 #' If not provided (default) the function searches into the current working directory.
 #' @param output Character string. Controls whether a complete data frame ('data.frame') 
@@ -46,13 +46,13 @@ match_wav_case <- function(X, path = NULL, output = "data.frame", verbose = TRUE
 
   #check path to working directory
   if (is.null(path)) path <- getwd() else 
-    if (!dir.exists(path)) stop("'path' provided does not exist") else
+    if (!dir.exists(path)) stop2("'path' provided does not exist") else
       path <- normalizePath(path)
     
 # list  wav files in path
 wvs <- list.files(path = path, pattern = "\\.wav$", ignore.case = TRUE)
 
-if (length(wvs) == 0) stop("no sound files are found in working directory")
+if (length(wvs) == 0) stop2("no sound files are found in working directory")
 
 # remove extension for both names in lbh_selec_table and wavs
 sf.no.ext <- gsub(".wav", "", X$sound.files, ignore.case = TRUE)