+++ R/raven_batch_detec.R
@@ -21,7 +21,7 @@
 #' @param view.preset Character string specifying the name of the window preset to be used. Not require for 'Amplitude Detector' (see 'detector.type' argument).
 #' If \code{NULL} (default) then the 'Default' window preset is used. 
 #' @param relabel_colms Logical. If  \code{TRUE} (default) columns are labeled to 
-#' match the selection table format from the acoustic analysis package \code{\link{warbleR}}
+#' match the selection table format from the acoustic analysis package \href{https://cran.r-project.org/package=warbleR}{warbleR}
 #' @param pb Logical argument to control progress bar. Default is \code{TRUE}.
 #' @param parallel Numeric. Controls whether parallel computing is applied.
 #' It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).
@@ -65,12 +65,8 @@ raven_batch_detec <- function(raven.path = NULL, sound.files, path = NULL, detec
 {
   
   # check path to working directory
-  if (is.null(path)) path <- getwd() else if (!dir.exists(path)) stop("'path' provided does not exist") else
+  if (is.null(path)) path <- getwd() else if (!dir.exists(path)) stop2("'path' provided does not exist") else
     path <- normalizePath(path)
-  
-  # set progress bar back to original
-  on.exit(pbapply::pboptions(type = .Options$pboptions$type), 
-          add = TRUE)
   
   # reset working directory 
   wd <- getwd()
@@ -79,19 +75,19 @@ raven_batch_detec <- function(raven.path = NULL, sound.files, path = NULL, detec
   
   # check path
   if (is.null(raven.path))
-    stop("Path to 'Raven' folder must be provided")  else
-      if (!dir.exists(raven.path)) stop("'raven.path' provided does not exist")
+    stop2("Path to 'Raven' folder must be provided")  else
+      if (!dir.exists(raven.path)) stop2("'raven.path' provided does not exist")
   
   setwd(raven.path)
   
   # check detector type
-  if (!detector.type %in% c("Amplitude Detector", "Band Limited Energy Detector", "Band Limited Entropy Detector")) stop("'detector.type' not recognized")
+  if (!detector.type %in% c("Amplitude Detector", "Band Limited Energy Detector", "Band Limited Entropy Detector")) stop2("'detector.type' not recognized")
   
   # check detector name
-  if (!file.exists(file.path(raven.path, "Presets/Detector", detector.type, detector.preset))) stop("'detector.preset' file not found")
+  if (!file.exists(file.path(raven.path, "Presets/Detector", detector.type, detector.preset))) stop2("'detector.preset' file not found")
   
   # check view  preset name
-  # if (!file.exists(file.path(raven.path, "Presets/Sound Window", view.preset))) stop("'view.preset' file not found")
+  # if (!file.exists(file.path(raven.path, "Presets/Sound Window", view.preset))) stop2("'view.preset' file not found")
   
   sf <- sound.files <- as.character(sound.files)
   
@@ -101,7 +97,7 @@ raven_batch_detec <- function(raven.path = NULL, sound.files, path = NULL, detec
   #count number of sound files in working directory and if 0 stop
   sound.files <- sound.files[sound.files %in% recs.wd]
   if (length(sound.files) == 0)
-    stop("The sound files are not in the working directory")
+    stop2("The sound files are not in the working directory")
   
   # remove sound files not found
   if (length(sound.files) != length(sf)) 
@@ -114,14 +110,11 @@ raven_batch_detec <- function(raven.path = NULL, sound.files, path = NULL, detec
   # check if raven executable is "Raven" or "RavenPro" (changed in Raven Pro 1.6)
   rav.exe <- list.files(path = raven.path, pattern =  "Raven$|Raven.app$|Raven.exe$|Raven\ Pro$|Raven\ Pro.app$|Raven\ Pro.exe$")
   
-  
-  if (pb) pbapply::pboptions(type = "timer") else pbapply::pboptions(type = "none")
-  
   # set clusters for windows OS
   if (Sys.info()[1] == "Windows" & parallel > 1)
     cl <- parallel::makePSOCKcluster(getOption("cl.cores", parallel)) else cl <- parallel  
   
-  out <- pbapply::pblapply(sound.files,  cl = cl, function(x) {
+  out <- warbleR:::.pblapply(pbar = pb, X = sound.files, cl = cl, message = "running raven batch detection", total = 1, function(x) {
     
     # view and detector preset together to fix it when view preset not need it  
     view.detector <- if (detector.type == "Amplitude Detector")  paste0("-detPreset:", detector.preset) else