+++ R/run_raven.R
@@ -64,7 +64,7 @@ run_raven <- function(raven.path = NULL, sound.files = NULL, path = NULL, at.the
   
   #check path to working directory
   if (is.null(path)) path <- getwd() else 
-    if (!dir.exists(path)) stop("'path' provided does not exist") else
+    if (!dir.exists(path)) stop2("'path' provided does not exist") else
       path <- normalizePath(path)
   
     # return to current wd on exit
@@ -73,18 +73,13 @@ run_raven <- function(raven.path = NULL, sound.files = NULL, path = NULL, at.the
     
     
   if (is.null(raven.path))
-    stop("Path to 'Raven' folder must be provided")  else
-      if (!dir.exists(raven.path)) stop("'raven.path' provided does not exist") else 
+    stop2("Path to 'Raven' folder must be provided")  else
+      if (!dir.exists(raven.path)) stop2("'raven.path' provided does not exist") else 
         setwd(raven.path)
-    
-  # set progress bar back to original
-  on.exit(pbapply::pboptions(type = .Options$pboptions$type), 
-          add = TRUE)
-  
    
   if (!is.null(view.preset))
    {
-    if (!any(view.preset %in% list.files(path = file.path(raven.path, "Presets/Sound Window")))) stop("'view.preset' provided not found")
+    if (!any(view.preset %in% list.files(path = file.path(raven.path, "Presets/Sound Window")))) stop2("'view.preset' provided not found")
     
   def.view.p <- grep("^Default", value = TRUE,  list.files(path = file.path(raven.path, "Presets/Sound Window")))
 
@@ -120,7 +115,7 @@ on.exit(unlink(file.path(raven.path, "Presets/Sound Window", grep("^temp.Default
   #count number of sound files in working directory and if 0 stop
   sound.files <- sound.files[sound.files %in% recs.wd]
   if (length(sound.files) == 0)
-    stop("The sound files are not in the working directory")
+    stop2("The sound files are not in the working directory")
   
   # remove sound files not found
   if (length(sound.files) != length(sf)) 
@@ -138,7 +133,7 @@ on.exit(unlink(file.path(raven.path, "Presets/Sound Window", grep("^temp.Default
     if (length(sound.files) == 0) 
     {
       cat("All sound files have a selection table in Raven's selection folder")
-      stop("")}
+      stop2("")}
     }
    
     # check if sound file names contains directory and fix
@@ -151,13 +146,11 @@ on.exit(unlink(file.path(raven.path, "Presets/Sound Window", grep("^temp.Default
   # subset by groups of sound files according to at the time
   sq <- unique(c(seq(1, length(sound.files), by = at.the.time)))
   
-  if (pb) pbapply::pboptions(type = "timer") else pbapply::pboptions(type = "none")
-  
   # check if raven executable is "Raven" or "RavenPro" (changed in Raven Pro 1.6)
   rav.exe <- list.files(path = raven.path, pattern =  "Raven$|Raven.app$|Raven.exe$|Raven\ Pro$|Raven\ Pro.app$|Raven\ Pro.exe$")
   
   # run loop over files
-  out <- pbapply::pblapply(sq, function(x)
+  out <- warbleR:::.pblapply(pbar = pb, X = sq, message = "running raven", total = 1, function(x)
     {
  
     fls <- sound.files[x:(x + at.the.time - 1)]