+++ R/to_sound_selection.R
@@ -54,16 +54,16 @@ to_sound_selection <- function(path = NULL, dest.path = NULL, recursive = FALSE,
   
   #check path to working directory
   if (is.null(path)) path <- getwd() else 
-    if (!dir.exists(path)) stop("'path' provided does not exist") else
+    if (!dir.exists(path)) stop2("'path' provided does not exist") else
       path <- normalizePath(path)
  
   # check dest.path
   if (is.null(dest.path)) dest.path <- path else 
-    if (!dir.exists(dest.path)) stop("'path' provided does not exist") 
+    if (!dir.exists(dest.path)) stop2("'path' provided does not exist") 
 
   sel.txt <- list.files(pattern = ".txt$", full.names = TRUE, recursive = recursive, ignore.case = TRUE, path = path)
 
-  if (length(sel.txt) == 0) stop("No selection .txt files in working directory/'path' provided")
+  if (length(sel.txt) == 0) stop2("No selection .txt files in working directory/'path' provided")
   
   options(warn = -1)
   
@@ -77,7 +77,7 @@ to_sound_selection <- function(path = NULL, dest.path = NULL, recursive = FALSE,
   read_sels2_FUN <- function(i, sel.txt)
   {  
     a <- try(utils::read.table(sel.txt[i], header = TRUE, sep = "\t", fill = TRUE, stringsAsFactors = FALSE, check.names = FALSE), silent = TRUE)
-    if (class(a) != "try-error")
+    if (!methods::is(a, "try-error"))
       {
     
     rf <- readLines(sel.txt[i])  
@@ -113,14 +113,11 @@ to_sound_selection <- function(path = NULL, dest.path = NULL, recursive = FALSE,
     }
 }
     
-  # set pb options 
-  pbapply::pboptions(type = ifelse(pb, "timer", "none"))
-  
   # set clusters for windows OS
   if (Sys.info()[1] == "Windows" & parallel > 1)
     cl <- parallel::makePSOCKcluster(getOption("cl.cores", parallel)) else cl <- parallel  
   
-  out <- pbapply::pblapply(seq_len(length(sel.txt)), cl = cl, function(i) {
+  out <- warbleR:::.pblapply(pbar = pb, X = seq_len(length(sel.txt)), cl = cl, message = "converting into sound selection table", total = 1, function(i) {
     read_sels2_FUN(i, sel.txt = sel.txt)
   })